Changed index type to size_t for StringTokenizer
Added missing header file to utils.h

+++ src/StringTokenizer.h
@@ -112,7 +112,7 @@ namespace utils{
     std::string delim;
     std::string string;
     bool keepDelim;
-    ushort index;
+    size_t index;
 
   };
 }
